Developer-App/requirements.txt
kivy>=2.2.0
kivymd>=1.1.1
plyer>=2.1.0
sqlite3
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
